restructure component directory
