--- src/components/CommunityLists/Community.js
@@ -1,12 +1,0 @@
-import React from "react"
-import "./style.css"
-
-const Community = () => {
-  return (
-    <div className="community">
-      <div />
-    </div>
-  )
-}
-
-export default Community
