--- src/components/CommunityLists/CommunityList.js
@@ -1,13 +1,0 @@
-import React from "react"
-import "./style.css"
-
-// container for a list of communities, used in Search and UserProfile
-const CommunityList = () => {
-  return (
-    <div className="communitylist">
-      <div />
-    </div>
-  )
-}
-
-export default CommunityList
