--- src/components/CommunityLists/style.css
@@ -1,1 +1,0 @@
-/* CSS for listing out communities, used in Search and UserProfile */
