--- src/components/LogIn/Login.js
@@ -1,12 +1,0 @@
-import React from "react"
-import "./style.css"
-
-const Login = () => {
-  return (
-    <div className="login">
-      <div />
-    </div>
-  )
-}
-
-export default Login
