--- src/components/LogIn/style.css
@@ -1,1 +1,0 @@
-/* CSS for the Login Page */
