+++ src/components/NavBar/NavBar.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { Link } from "react-router-dom"
-import "./style.css"
+import "./NavBar.css"
 import { Typography } from "antd"
 import ACMLogo from "../../images/acm-logo.png"
