+++ src/components/UserList/UserList.css
@@ -1,0 +1,19 @@
+.contact-list-container {
+    padding: 10px 10px 10px 10px;
+    margin-bottom: 50px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.contact-list-edit-button-container {
+    padding: 10px 0 10px 10px;
+}
+
+.contact-list-contacts-container {
+    box-shadow: 0px 0px 7px 1px rgba(0, 0, 0, 0.1);
+}
+
+.contact-list-edit-button {
+    box-shadow: 0px 0px 7px 1px rgba(0, 0, 0, 0.1);
+}
