+++ src/pages/user-profile-page.js
@@ -3,10 +3,10 @@ import React, { useState } from "react"
 import { Tabs, Divider } from "antd"
 import "./user-profile-page.css"
 import "antd/dist/antd.css"
-import UserHeader from "../components/UserProfile/UserHeader"
-import ContactList from "../components/UserProfile/ContactList"
-import UserUpdateToolbar from "../components/UserProfile/UserUpdateToolbar"
-import UserList from "../components/UserProfile/UserList"
+import UserHeader from "../components/UserHeader/UserHeader"
+import ContactList from "../components/ContactList/ContactList"
+import UserUpdateToolbar from "../components/UI/UserUpdateToolbar"
+import UserList from "../components/UserList/UserList"
 
 const { TabPane } = Tabs
